xenserver: Set network-uuid for internal bridges
The "citrix" branch contains a script called dump-vif-details that would
write detailed information about the vifs into ovs-vswitchd.conf.
Additionally, it would add the network-uuid information for internal
bridges, since the interface-reconfigure script that normally takes care
of this is not called on internal bridges.  The Midnight Ride release
no longer needs to call dump-vif-details, but the functionality to write
the network-uuid for internal bridges was not put elsewhere.  This
commit addresses that.

It introduces the vif-on-internal-bridge script, which is a hack until a
more appropriate method, such XAPI writing out the information to
xenstore, becomes available.

Bug #2208