ofproto: Use active time for Flow Expiration duration (OpenFlow 0.9)
In pre-0.9 versions of OpenFlow, the duration field in the flow
expiration message was the amount of time that the flow was in a table.
If a long idle timeout was set, it could be difficult to determine how
long the flow was actually active.  In OpenFlow 0.9, this was changed to
only be the length of time that the flow was active.

NOTE: OVS at this point is not wire-compatible with OpenFlow 0.9 until the
final commit in this OpenFlow 0.9 set.