xenserver: Use ovs-appctl to reopen log files
During the changes to use the configuration database the "reload"
command was removed from the init script.  In addition to reloading the
configuration file, binaries also reopened their log files, which is the
behavior the logrotate configuration expected.  This change makes the
logrotate configuration no longer user the "reload" command.