Add ability to set controller for specific bridges
Some applications would prefer that the controller only be enabled for
a subset of bridges.  This commit introduces an alternative syntax for
the XAPI vSwitchController key to enable the specification of such a
configuration.

The implementation of the configuration is currently a hack.  The
problem is that some bridges (especially internal bridges) are only
created when a VIF is added to them.  This means that attempting to
set the controller on boot as we used to do won't work.  Ideally we
would should hook the creation of the bridge and set the controller at
that time.  However, in XenServer 5.5 I can't find an appropriate
place to hook to do so.

To meet immediate requirements, this commit hacks the startup-based
approach by spawning background processes that watch for the creation
of the bridge and set the controller when it becomes available.  We'll
have to replace it with a better and more robust solution ASAP.