rconn: Decentralize and improve disconnection logging.
In some cases the rconn library was logging disconnections twice, and in
other cases it was not logging them at all.  This cleans it up, so that
each disconnection or connection failure should cause one log message.

This commit makes disconnect()'s 'error' argument unused, but it will be
used again in the following commit.